+++ src/menu_systems/text_input.rs
@@ -236,7 +236,7 @@ impl InterfaceItem for TextInput {
             text_buffer.change_cursor_bg_color(self.bg_color_unfocused);
             text_buffer.change_cursor_fg_color(self.fg_color_unfocused);
         }
-        text_buffer.move_cursor(self.base.x as i32, self.base.y as i32);
+        text_buffer.move_cursor(self.base.x, self.base.y);
 
         let text_w_offset: u32;
         if self.base.is_focused() && self.caret != 0.0 {
